API/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
]
